CommandLine refactoring step 1:  Move Run.exe version into common and take out usages of Setup object.
