+++ src/common/CommandLine.cs
@@ -7,6 +7,7 @@ using System.Diagnostics;
 using System.Text;
 using System.Collections.Generic;
 using System.IO;
+using Microsoft.DotNet.Execute;
 
 using System.Reflection;
 
@@ -19,8 +20,8 @@ namespace Microsoft.Fx.CommandLine
     /// The code:CommandLineParser is a utility for parsing command lines. Command lines consist of three basic
     /// entities. A command can have any (or none) of the following (separated by whitespace of any size).
     ///
-    ///    * PARAMETERS - this are non-space strings. They are positional (logicaly they are numbered). Strings
-    ///        with space can be specified by enclosing in double quotes.
+    ///    * PARAMETERS - these are non-space strings. They are positional (logicaly they are numbered). Strings
+    ///        with whitespace can be specified by enclosing in double quotes.
     ///
     ///    * QUALIFIERS - Qualifiers are name-value pairs. The following syntax is supported.
     ///        * -QUALIFIER
@@ -29,8 +30,8 @@ namespace Microsoft.Fx.CommandLine
     ///        * -QUALIFIER VALUE
     ///
     ///      The end of a value is delimited by space. Again values with spaces can be encoded by enclosing them
-    ///      the value (or the whole qualifer-value string), in double quotes. The first form (where a value is
-    ///      not specified is only available for boolean qualifiers, and boolean values can not use the form where
+    ///      within the value (or the whole qualifer-value string), in double quotes. The first form (where a value is
+    ///      not specified) is only available for boolean qualifiers, and boolean values can not use the form where
     ///      the qualifer and value are separated by space. The '/' character can also be used instead of the '-'
     ///      to begin a qualifier.
     ///
@@ -50,22 +51,24 @@ namespace Microsoft.Fx.CommandLine
     /// independent of one another. For example a for example For example a program might have 'checkin'
     /// 'checkout' 'list' commands, and each of these commands has a different set of parameters that are needed
     /// and qualifiers that are allowed. (for example checkout will take a list of file names, list needs nothing,
-    /// and checkin needs a comment). Additionally some qualifiers (like say -dataBaseName can apply to any of hte
-    /// commands). Thus You would like to say that the following command lines are legal
+    /// and checkin needs a comment). Additionally some qualifiers (like say -dataBaseName can apply to any of the
+    /// commands). 
+    /// 
+    /// The following command lines are legal:
     ///
     ///     * EXE -checkout MyFile1 MyFile -dataBaseName:MyDatabase
     ///     * EXE -dataBaseName:MyDatabase -list
     ///     * EXE -comment "Specifying the comment first" -checkin
     ///     * EXE -checkin -comment "Specifying the comment afterward"
     ///
-    /// But the following are not
+    /// But the following are not:
     ///
     ///     * EXE -checkout
     ///     * EXE -checkout -comment "hello"
     ///     * EXE -list MyFile
     ///
-    /// You do this by specifying 'checkout', 'list' and 'checkin' as parameters sets. On the command line they
-    /// look like boolean qualifiers, however they have additional sematantics. They must come before any
+    /// You do this by specifying 'checkout', 'list' and 'checkin' as parameter sets. On the command line they
+    /// look like boolean qualifiers, however they have additional semantics. They must come before any
     /// positional parameters (because they affect whether the parameters are allowed and what they are named),
     /// and they are mutually exclusive. Each parameter set gets its own set of parameter definitions, and
     /// qualifiers can either be associated with a particular parameter set (like -comment) or global to all
@@ -80,7 +83,7 @@ namespace Microsoft.Fx.CommandLine
     ///
     /// #DefaultParameterSet
     ///
-    /// One parameters set (which has the empty string name), is special in that it is used when no other
+    /// One parameter set (which has the empty string name), is special in that it is used when no other
     /// parameter set is matched. This is the default parameter set. For example, if -checkout was defined to be
     /// the default parameter set, then the following would be legal.
     ///
@@ -96,7 +99,7 @@ namespace Microsoft.Fx.CommandLine
     /// Because whitespace can separate a qualifier from its value AND Qualifier from each other, and because
     /// parameter values might start with a dash (and thus look like qualifiers), the syntax is ambiguous. It is
     /// disambigutated with the following rules.
-    ///     * The command line is parsed into 'arguments' that are spearated by whitespace. Any string enclosed
+    ///     * The command line is parsed into 'arguments' that are separated by whitespace. Any string enclosed
     ///         in "" will be a single argument even if it has embedded whitespace. Double quote characters can
     ///         be specified by \" (and a \" literal can be specified by \\" etc).
     ///     * Arguments are parsed into qualifiers. This parsing stops if a '--' argument is found. Thus all
@@ -154,18 +157,18 @@ namespace Microsoft.Fx.CommandLine
 #endif
     /// #RequiredAndOptional
     ///
-    /// Parameters and qualifiers can be specified as required (the default), or optional. Makeing the default
-    /// required was choses to make any mistakes 'obvious' since the parser will fail if a required parameter is
+    /// Parameters and qualifiers can be specified as required (the default), or optional. Making the default
+    /// required was chosen to make any mistakes 'obvious' since the parser will fail if a required parameter is
     /// not present (if the default was optional, it would be easy to make what should have been a required
     /// qualifier optional, leading to business logic failiure).
     ///
     /// #ParsedValues
     ///
-    /// The class was designed maximize programmer convinience. For each parameter, only one call is needed to
+    /// The class was designed maximize programmer convenience. For each parameter, only one call is needed to
     /// both define the parameter, its help message, and retrive its (strong typed) value. For example
     ///
     ///      * int count = 5;
-    ///      * parser.DefineOptionalQualifier("count", ref count, "help for optional debugs qualifier");
+    ///      * parser.DefineOptionalQualifier("count", ref count, "help for optional count qualifier");
     ///
     /// Defines a qualifier 'count' and will place its value in the local variable 'count' as a integer. Default
     /// values are supported by doing nothing, so in the example above the default value will be 5.
@@ -174,10 +177,10 @@ namespace Microsoft.Fx.CommandLine
     /// string argument and returning that type. This is true for all primtive types, DateTime, Enumerations, and
     /// any user defined type that follows this convention.
     ///
-    /// Array types: The parser has special knowedge of arrays. If the type of a qualifier is an array, then the
+    /// Array types: The parser has special knowledge of arrays. If the type of a qualifier is an array, then the
     /// string value is assumed to be a ',' separated list of strings which will be parsed as the element type of
     /// the array. In addition to the ',' syntax, it is also legal to specify the qualifier more than once. For
-    /// example given the defintion
+    /// example given the definition
     ///
     ///      * int[] counts;
     ///      * parser.DefineOptionalQualifier("counts", ref counts, "help for optional counts qualifier");
@@ -231,7 +234,7 @@ namespace Microsoft.Fx.CommandLine
     /// Since this is stuff that all applications will likely need the
     /// code:CommandLineParser.ParseForConsoleApplication was created to do all of this for you, thus making it
     /// super-easy to make a production quality parser (and concentrate on getting your application logic instead
-    /// of command line parsing. Here is an example which defines a 'Ping' command. If you will notice there are
+    /// of command line parsing). Here is an example which defines a 'Ping' command. If you will notice there are
     /// very few lines of code that are not expressing something very specific to this applications. This is how
     /// it should be!
 #if EXAMPLE2
@@ -262,7 +265,7 @@ namespace Microsoft.Fx.CommandLine
     /// don't need to be passed around to many routines.  In general, however it is often a better idea to
     /// create a class whose sole purpose is to act as a repository for the parsed arguments.   This also nicely
     /// separates all command line processing into a single class.   This is how the ping example would look  in
-    /// that style. Notice that the main program no longer holds any command line processing logic.  and that
+    /// that style. Notice that the main program no longer holds any command line processing logic, and that
     /// 'commandLine' can be passed to other routines in bulk easily.
 #if EXAMPLE3
 class CommandLineParserExample3
@@ -297,37 +300,43 @@ class CommandLine
     {
         /// <summary>
         /// If you are building a console Application, there is a common structure to parsing arguments. You want
-        /// the text formated and output for console windows, and you want /? to be wired up to do this. All
+        /// the text formatted and output for console windows, and you want /? to be wired up to do this. All
         /// errors should be caught and displayed in a nice way.  This routine does this 'boiler plate'.
         /// </summary>
         /// <param name="parseBody">parseBody is the body of the parsing that this outer shell does not provide.
         /// in this delegate, you should be defining all the command line parameters using calls to Define* methods.
         ///  </param>
-        public static bool ParseForConsoleApplication(Action<CommandLineParser> parseBody, string[] args)
+        public static bool ParseForConsoleApplication(Action<CommandLineParser> parseBody, string[] args, Setup setupContent)
         {
             return Parse(parseBody, parser =>
             {
                 var parameterSetTofocusOn = parser.HelpRequested;
+                string helpString;
                 if (parameterSetTofocusOn.Length == 0)
                 {
                     parameterSetTofocusOn = null;
+                    //helpString = parser.GetHelp(GetConsoleWidth() - 1, parameterSetTofocusOn, true);
                 }
 
-                string helpString = parser.GetHelp(GetConsoleWidth() - 1, parameterSetTofocusOn, true);
+                helpString = parser.GetIntroTextForHelp(GetConsoleWidth() - 1).ToString();
+                helpString += parser.GetHelp(GetConsoleWidth() - 1, parameterSetTofocusOn, true);
                 DisplayStringToConsole(helpString);
             }, (parser, ex) =>
             {
-                Console.WriteLine("Error: " + ex.Message);
-                Console.WriteLine("Use -? for help.");
+                Console.Error.WriteLine("Error: " + ex.Message);
+                
+                Console.ForegroundColor = ConsoleColor.DarkYellow;
+                Console.WriteLine("Our dev workflow has changed! Use -? for help in the new options we have and how to pass parameters now.");
+                Console.ResetColor();
             },
-            args);
+            setupContent, args);
         }
 
-        public static bool Parse(Action<CommandLineParser> parseBody, Action<CommandLineParser> helpHandler, Action<CommandLineParser, Exception> errorHandler, string[] args)
+        public static bool Parse(Action<CommandLineParser> parseBody, Action<CommandLineParser> helpHandler, Action<CommandLineParser, Exception> errorHandler, Setup setupContent, string[] args)
         {
             var help = false;
-            CommandLineParser parser = new CommandLineParser(args);
-            parser.DefineOptionalQualifier("?", ref help, "Print this help guide.");
+            CommandLineParser parser = new CommandLineParser(args, setupContent);
+            parser.DefineOptionalQualifier("?", ref help, "Print this help guide.", null, null);
 
             try
             {
@@ -374,9 +383,9 @@ class CommandLine
         /// <param name="helpText">Text to print for this qualifier.  It will be word-wrapped.  Newlines indicate
         /// new paragraphs.</param>
         /// </summary>
-        public void DefineOptionalQualifier<T>(string name, ref T retVal, string helpText)
+        public void DefineOptionalQualifier<T>(string name, ref T retVal, string helpText, string defaultValue, List<string> legalValues)
         {
-            object obj = DefineQualifier(name, typeof(T), retVal, helpText, false);
+            object obj = DefineQualifier(name, typeof(T), retVal, helpText, false, defaultValue, legalValues);
             if (obj != null)
                 retVal = (T)obj;
         }
@@ -387,9 +396,9 @@ class CommandLine
         /// <param name="helpText">Text to print for this qualifer.  It will be word-wrapped.  Newlines indicate
         /// new paragraphs.</param>
         /// </summary>
-        public void DefineQualifier<T>(string name, ref T retVal, string helpText)
+        public void DefineQualifier<T>(string name, ref T retVal, string helpText, string defaultValue, List<string> legalValues)
         {
-            object obj = DefineQualifier(name, typeof(T), retVal, helpText, true);
+            object obj = DefineQualifier(name, typeof(T), retVal, helpText, true, defaultValue, legalValues);
             if (obj != null)
                 retVal = (T)obj;
         }
@@ -567,6 +576,7 @@ class CommandLine
             {
                 if (_qualifiersUseOnlyDash != value)
                     ThrowIfNotFirst("OnlyDashForQualifiers");
+
                 _qualifiersUseOnlyDash = value;
             }
         }
@@ -584,32 +594,20 @@ class CommandLine
             set { _lastQualifierWins = value; }
         }
 
+        public static string ExtraParameters
+        {
+            get { return _extraParameters; }
+        }
+
         // These routines are typically are not needed because ParseArgsForConsoleApp does the work.
-#if !COREFX
-        public CommandLineParser() : this(Environment.CommandLine) { }
-#endif
         public CommandLineParser(string commandLine)
         {
-            _args = new List<string>();
             ParseWords(commandLine);
         }
-        public CommandLineParser(string[] args)
+        public CommandLineParser(string[] args, Setup setupContent)
         {
-            _args = new List<string>();
-            _seenExeArg = true;
-
-            for (int i = 0; i < args.Length; i++)
-            {
-                if (args[i][0] == '@')
-                {
-                    // Parse response files
-                    ParseWords(args[i]);
-                }
-                else
-                {
-                    _args.Add(args[i]);
-                }
-            }
+            _args = new List<string>(args);
+            _setupContent = setupContent;
         }
 
         /// <summary>
@@ -628,11 +626,20 @@ class CommandLine
             }
 
             // Find any 'unused' parameters;
-            while (_curPosition < _args.Count && _args[_curPosition] == null)
-                _curPosition++;
-
-            if (_curPosition < _args.Count)
-                throw new CommandLineParserException("Extra positional parameter: " + _args[_curPosition] + ".");
+            List<string> unusedParameters = new List<string>();
+            foreach(string param in _args)
+            {
+                if(!string.IsNullOrEmpty(param))
+                {
+                    unusedParameters.Add(param);
+                }
+            }
+            
+            if (unusedParameters.Count > 0)
+            {
+                throw new CommandLineParserException("Parameter not recognized: " + unusedParameters[0] + ".");
+            }
+                
 
             // TODO we should null out data structures we no longer need, to save space.
             // Not critical because in the common case, the parser as a whole becomes dead.
@@ -702,12 +709,15 @@ class CommandLine
 
             // Create the 'Usage' line;
             string appName = GetEntryAssemblyName();
+            string command = parameterSetParameter.Syntax();
+            sb.Append(command).AppendLine().AppendLine();
 
             sb.Append("Usage: ").Append(appName);
             if (parameterSetName.Length > 0)
             {
                 sb.Append(' ');
-                sb.Append(parameterSetParameter.Syntax(false, false));
+                sb.Append(command);
+                sb.Append(" [Action] (global settings)");
             }
 
             bool hasQualifiers = false;
@@ -721,53 +731,71 @@ class CommandLine
                 if (parameter.IsPositional)
                 {
                     hasParameters = true;
-                    sb.Append(' ').Append(parameter.Syntax(false, false));
+                    sb.Append(' ').Append(parameter.Syntax());
                 }
                 else
+                {
                     hasQualifiers = true;
+                    break;
+                }
             }
             sb.AppendLine();
 
             // Print the help for the command itself.
-            if (parameterSetParameter != null && !string.IsNullOrEmpty(parameterSetParameter.HelpText))
+            /*if (parameterSetParameter != null && !string.IsNullOrEmpty(parameterSetParameter.HelpText))
             {
                 sb.AppendLine();
-                if (parameterSetParameter != null && parameterSetParameter.HelpText != null)
-                    Wrap(sb.Append("  "), parameterSetParameter.HelpText, 2, "  ", maxLineWidth);
-            }
+                Wrap(sb.Append("  "), parameterSetParameter.HelpText, 2, "  ", maxLineWidth);
+            }*/
 
             if (hasParameters)
             {
-                sb.AppendLine().Append("  Parameters:").AppendLine();
+                sb.AppendLine().Append("Parameters:").AppendLine();
                 for (int i = parameterSetBody; i < _parameterDescriptions.Count; i++)
                 {
                     CommandLineParameter parameter = _parameterDescriptions[i];
                     if (parameter.IsParameterSet)
                         break;
                     if (parameter.IsPositional)
+                    {
                         ParameterHelp(parameter, sb, QualifierSyntaxWidth, maxLineWidth);
+                    }
                 }
             }
-
-            string globalQualifiers = null;
-            if (displayGlobalQualifiers)
-                globalQualifiers = GetHelpGlobalQualifiers(maxLineWidth);
-
-            if (hasQualifiers || !string.IsNullOrEmpty(globalQualifiers))
+            else if (hasQualifiers)
             {
-                sb.AppendLine().Append("  Qualifiers:").AppendLine();
+                sb.AppendLine().Append("Actions:").AppendLine();
+
                 for (int i = parameterSetBody; i < _parameterDescriptions.Count; i++)
                 {
                     CommandLineParameter parameter = _parameterDescriptions[i];
                     if (parameter.IsParameterSet)
                         break;
                     if (parameter.IsNamed)
+                    {
                         ParameterHelp(parameter, sb, QualifierSyntaxWidth, maxLineWidth);
+                        string commandSettingsHelp = _setupContent.GetHelpCommand(parameterSetParameter.Name, parameter.Name);
+                        Wrap(sb, commandSettingsHelp, QualifierSyntaxWidth, new string(' ', QualifierSyntaxWidth), maxLineWidth, false);
+                    }
                 }
-                if (globalQualifiers != null)
-                    sb.Append(globalQualifiers);
+            }
+            else
+            {
+                string commandSettingsHelp = _setupContent.GetHelpCommand(parameterSetParameter.Name);
+                Wrap(sb, commandSettingsHelp, QualifierSyntaxWidth, new string(' ', QualifierSyntaxWidth), maxLineWidth, false);
             }
 
+            string globalQualifiers = null;
+            if (displayGlobalQualifiers)
+                globalQualifiers = GetHelpGlobalQualifiers(maxLineWidth);
+            if(!string.IsNullOrEmpty(globalQualifiers))
+            {
+                sb.AppendLine().Append('-', maxLineWidth - 1).AppendLine();
+                sb.Append("Global Settings:").AppendLine();
+                sb.AppendLine();
+                sb.Append(globalQualifiers);
+            }
+                
             return sb.ToString();
         }
         /// <summary>
@@ -789,7 +817,9 @@ class CommandLine
         {
             public string Name { get { return _name; } }
             public Type Type { get { return _type; } }
-            public object DefaultValue { get { return _defaultValue; } }
+            public string DefaultValue { get { return _defaultValue; } }
+            public List<string> LegalValues { get { return _legalValues; } }
+            public object Value { get { return _value; } }
             public bool IsRequired { get { return _isRequired; } }
             public bool IsPositional { get { return _isPositional; } }
             public bool IsNamed { get { return !_isPositional; } }
@@ -800,48 +830,16 @@ class CommandLine
                 return "<CommandLineParameter " +
                     "Name=\"" + _name + "\" " +
                     "Type=\"" + _type + "\" " +
+                    "DefaultValue=\"" + _defaultValue + "\" " +
                     "IsRequired=\"" + IsRequired + "\" " +
                     "IsPositional=\"" + IsPositional + "\" " +
                     "HelpText=\"" + HelpText + "\"/>";
             }
-            public string Syntax(bool printType, bool printDefaultValue)
+            public string Syntax()
             {
                 string ret = _name;
                 if (IsNamed)
                     ret = "-" + ret;
-                if (printType)
-                {
-                    // We print out arrays with the ... notiation, so we don't want the [] when we display the type
-                    Type displayType = Type;
-                    if (displayType.IsArray)
-                        displayType = displayType.GetElementType();
-                    if (displayType.GetTypeInfo().IsGenericType && displayType.GetGenericTypeDefinition() == typeof(Nullable<>))
-                        displayType = displayType.GetGenericArguments()[0];
-
-                    bool shouldPrint = true;
-                    // Bool type implied on named parameters
-                    if (IsNamed && _type == typeof(bool))
-                    {
-                        shouldPrint = false;
-                        if (_defaultValue != null && (bool)_defaultValue)
-                            shouldPrint = false;
-                    }
-                    // string type is implied on positional parameters
-                    if (IsPositional && displayType == typeof(string) && string.IsNullOrEmpty(_defaultValue as string))
-                        shouldPrint = false;
-
-                    if (shouldPrint)
-                    {
-                        ret = ret + (IsNamed ? ":" : ".") + displayType.Name.ToUpper();
-                        if (printDefaultValue && _defaultValue != null)
-                        {
-                            string defValue = _defaultValue.ToString();
-                            if (defValue.Length < 40)   // TODO is this reasonable?
-                                ret += "(" + defValue + ")";
-                        }
-                    }
-                }
-
                 if (Type.IsArray)
                     ret = ret + (IsNamed ? "," : " ") + "...";
 
@@ -851,20 +849,31 @@ class CommandLine
             }
 
             #region private
-            internal CommandLineParameter(string Name, object defaultValue, string helpText, Type type,
-                bool isRequired, bool isPositional, bool isParameterSet)
+            internal CommandLineParameter(string Name, object value, string helpText, Type type,
+                bool isRequired, bool isPositional, bool isParameterSet, List<string> legalvalues, string defaultValue = null)
             {
                 _name = Name;
+                _value = value;
                 _defaultValue = defaultValue;
                 _type = type;
                 _helpText = helpText;
                 _isRequired = isRequired;
                 _isPositional = isPositional;
                 _isParameterSet = isParameterSet;
+                if(legalvalues != null)
+                {
+                    _legalValues = new List<string>(legalvalues);
+                }
+                else
+                {
+                    _legalValues = new List<string>();
+                }
             }
 
             private string _name;
-            private object _defaultValue;
+            private object _value;
+            private string _defaultValue;
+            private List<string> _legalValues;
             private string _helpText;
             private Type _type;
             private bool _isRequired;
@@ -924,6 +933,8 @@ class CommandLine
         /// <param name="commandLine"></param>
         private void ParseWords(string commandLine)
         {
+            // TODO review this carefully.
+            _args = new List<string>();
             int wordStartIndex = -1;
             bool hasExcapedQuotes = false;
             bool isResponseFile = false;
@@ -1090,7 +1101,7 @@ class CommandLine
                         }
                     }
                     if (nextPos >= 0)
-                        _dashedParameterEncodedPositions[name] = SetMulitple(nextPos);
+                        _dashedParameterEncodedPositions[name] = SetMultiple(nextPos);
                     else
                         _dashedParameterEncodedPositions.Remove(name);
                 }
@@ -1098,7 +1109,7 @@ class CommandLine
             return ret;
         }
 
-        // Phase 2 parsing works into things that look like qualifiers (but without  knowledge of which qualifiers the command supports)
+        // Phase 2 parsing works into things that look like qualifiers (but without knowledge of which qualifiers the command supports)
         /// <summary>
         /// Find the locations of all arguments that look like named parameters.
         /// </summary>
@@ -1107,6 +1118,7 @@ class CommandLine
             _dashedParameterEncodedPositions = new Dictionary<string, int>(StringComparer.OrdinalIgnoreCase);
 
             bool paramSetEncountered = false;
+            string parameterSet = string.Empty;
             for (int i = 0; i < _args.Count; i++)
             {
                 string arg = _args[i];
@@ -1116,27 +1128,43 @@ class CommandLine
                 string name = ParseParameterName(arg);
                 if (name != null)
                 {
-                    if (name.Length == 1 && IsDash(name[0]))
+                    if (IsDash(name[0]))
                     {
                         _args[i] = null;
+                        i++;
+                        string[] extraP = new string[_args.Count - i];
+                        for (int j = 0; i < _args.Count; i++, j++)
+                        {
+                            extraP[j] = _args[i];
+                            _args[i] = null;
+                        }
+                        _extraParameters = string.Join(" ", extraP);
                         break;
                     }
                     if (name == "?")        // Did the user request help?
                     {
                         _args[i] = null;
-                        _helpRequestedFor = String.Empty;
-                        if (i + 1 < _args.Count)
+                        if(paramSetEncountered)
+                        {
+                            _helpRequestedFor = parameterSet;
+                        }
+                        else
+                        {
+                            _helpRequestedFor = String.Empty;
+                        }
+                        
+                        /*if (i + 1 < _args.Count)
                         {
                             i++;
                             _helpRequestedFor = _args[i];
                             _args[i] = null;
-                        }
+                        }*/
                         _mustParseHelpStrings = true;
                         break;
                     }
                     int position = i;
                     if (_dashedParameterEncodedPositions.TryGetValue(name, out position))
-                        position = SetMulitple(position);
+                        position = SetMultiple(position);
                     else
                         position = i;
                     _dashedParameterEncodedPositions[name] = position;
@@ -1153,6 +1181,7 @@ class CommandLine
                         else if (IsParameterSetWithqualifiersMustBeFirst(String.Empty))        // Then we are the default parameter set
                             break;
                         paramSetEncountered = true;     // If we have hit a parameter, we must have hit a parameter set.
+                        parameterSet = arg;
                     }
                 }
             }
@@ -1170,12 +1199,12 @@ class CommandLine
             return false;
         }
 
-        // Phase 3, processing user defintions of qualifiers parameter sets etc.
+        // Phase 3, processing user defintions of qualifiers, parameter sets, etc.
         /// <summary>
         /// returns the index in the 'args' array of the next instance of the 'name' qualifier.   returns -1 if there is
         /// no next instance of the qualifer.
         /// </summary>
-        private object DefineQualifier(string name, Type type, object defaultValue, string helpText, bool isRequired)
+        private object DefineQualifier(string name, Type type, object qualifierValue, string helpText, bool isRequired, string defaultValue, List<string> legalValues)
         {
             Debug.Assert(_args != null);
             if (_dashedParameterEncodedPositions == null)
@@ -1183,18 +1212,18 @@ class CommandLine
 
             _qualiferEncountered = true;
             if (_mustParseHelpStrings)
-                AddHelp(new CommandLineParameter(name, defaultValue, helpText, type, isRequired, false, false));
+                AddHelp(new CommandLineParameter(name, qualifierValue, helpText, type, isRequired, false, false, legalValues, defaultValue));
             if (_skipDefinitions)
                 return null;
             if (_positionalArgEncountered && !_noSpaceOnQualifierValues)
-                throw new CommandLineParserDesignerException("Definitions of Named parameters must come before defintions of positional parameters");
+                throw new CommandLineParserDesignerException("Definitions of Named parameters must come before definitions of positional parameters");
 
             object ret = null;
             string[] alaises = null;
             if (_aliasDefinitions != null)
                 _aliasDefinitions.TryGetValue(name, out alaises);
 
-            int occuranceCount = 0;
+            int occurenceCount = 0;
             List<Array> arrayValues = null;
             for (;;)
             {
@@ -1219,7 +1248,7 @@ class CommandLine
                 }
                 else if (value == null)
                 {
-                    int valuePos = position + 1;
+                    /*int valuePos = position + 1;
                     if (_noSpaceOnQualifierValues || valuePos >= _args.Count || _args[valuePos] == null)
                     {
                         string message = "Parameter " + name + " is missing a value.";
@@ -1237,7 +1266,8 @@ class CommandLine
                     if (value.Length > 0 && IsQualifier(value))
                         throw new CommandLineParserException("Use " + name + ":" + value + " if " + value +
                             " is meant to be value rather than a named parameter");
-                    _args[valuePos] = null;
+                    _args[valuePos] = null;*/
+                    value = string.IsNullOrEmpty(defaultValue) ? "true" : defaultValue;
                 }
                 ret = ParseValue(value, type, name);
                 if (type.IsArray)
@@ -1247,19 +1277,19 @@ class CommandLine
                     arrayValues.Add((Array)ret);
                     ret = null;
                 }
-                else if (occuranceCount > 0 && !_lastQualifierWins)
+                else if (occurenceCount > 0 && !_lastQualifierWins)
                     throw new CommandLineParserException("Parameter " + name + " specified more than once.");
-                occuranceCount++;
+                occurenceCount++;
             }
 
-            if (occuranceCount == 0 && isRequired)
+            if (occurenceCount == 0 && isRequired)
                 throw new CommandLineParserException("Required named parameter " + name + " not present.");
 
             if (arrayValues != null)
-                ret = ConcatinateArrays(type, arrayValues);
+                ret = ConcatenateArrays(type, arrayValues);
             return ret;
         }
-        private object DefineParameter(string name, Type type, object defaultValue, string helpText, bool isRequired)
+        private object DefineParameter(string name, Type type, object parameterValue, string helpText, bool isRequired)
         {
             Debug.Assert(_args != null);
             if (_dashedParameterEncodedPositions == null)
@@ -1268,11 +1298,11 @@ class CommandLine
             if (!isRequired)
                 _optionalPositionalArgEncountered = true;
             else if (_optionalPositionalArgEncountered)
-                throw new CommandLineParserDesignerException("Optional positional parameters can't preceed required positional parameters");
+                throw new CommandLineParserDesignerException("Optional positional parameters can't precede required positional parameters");
 
             _positionalArgEncountered = true;
             if (_mustParseHelpStrings)
-                AddHelp(new CommandLineParameter(name, defaultValue, helpText, type, isRequired, true, false));
+                AddHelp(new CommandLineParameter(name, parameterValue, helpText, type, isRequired, true, false, null));
             if (_skipDefinitions)
                 return null;
 
@@ -1338,7 +1368,7 @@ class CommandLine
                 _defaultParamSetEncountered = true;
 
             if (_mustParseHelpStrings)
-                AddHelp(new CommandLineParameter(name, null, helpText, typeof(bool), true, _noDashOnParameterSets, true));
+                AddHelp(new CommandLineParameter(name, null, helpText, typeof(bool), true, _noDashOnParameterSets, true, null));
             if (_skipParameterSets)
                 return false;
 
@@ -1394,7 +1424,7 @@ class CommandLine
             return ret;
         }
 
-        private Array ConcatinateArrays(Type arrayType, List<Array> arrays)
+        private Array ConcatenateArrays(Type arrayType, List<Array> arrays)
         {
             int totalCount = 0;
             for (int i = 0; i < arrays.Count; i++)
@@ -1553,6 +1583,25 @@ class CommandLine
             }
         }
 
+        private StringBuilder GetIntroTextForHelp(int maxLineWidth)
+        {
+            string appName = GetEntryAssemblyName();
+            StringBuilder sb = new StringBuilder();
+            sb.AppendLine();
+            string text = "The Run Command Tool is now in charge of running the dev workflow steps. Each step has its own command and set of actions that are listed below.  " +
+                "You could also pass Global Settings to the commands.";
+            Wrap(sb, text, 0, String.Empty, maxLineWidth, true);
+            text = "To pass additional parameters that are not described in the Global Settings section, use `--`. After this command, the Run Command Tool will stop processing arguments and will pass all the information as it is to the selected command.";
+            Wrap(sb, text, 0, String.Empty, maxLineWidth, true);
+            text = "The information comes from a config.json file. By default the file is in the root of the repo. Otherwise the first parameter should be the path to the config.json file.";
+            Wrap(sb, text, 0, String.Empty, maxLineWidth, true);
+            text = "For more information about the Run Command Tool: https://github.com/dotnet/buildtools/blob/master/Documentation/RunCommand.md";
+            Wrap(sb, text, 0, String.Empty, maxLineWidth, true);
+            sb.AppendLine().AppendLine().Append("Syntax: run [Command] [Action] (global settings)");
+            sb.AppendLine().Append('-', maxLineWidth - 1).AppendLine();
+            return sb;
+        }
+
 
         /// <summary>
         /// Return a string giving the help for the command, word wrapped at 'maxLineWidth'
@@ -1562,24 +1611,22 @@ class CommandLine
             // Do we have non-default parameter sets?
             bool hasParamSets = false;
             foreach (CommandLineParameter parameter in _parameterDescriptions)
+            {
                 if (parameter.IsParameterSet && parameter.Name != String.Empty)
+                {
                     hasParamSets = true;
-
+                    break;
+                }
+            }
+            
             if (!hasParamSets)
                 return GetHelp(maxLineWidth, String.Empty, true);
+
             StringBuilder sb = new StringBuilder();
-
-
-            string appName = GetEntryAssemblyName();
-            string intro = "The " + appName + " application has a number of commands associated with it, " +
-                "each with its own set of parameters and qualifiers.  They are listed below.  " +
-                "Options that are common to all commands are listed at the end.";
-            Wrap(sb, intro, 0, String.Empty, maxLineWidth);
-
+                        
             // Always print the default parameter set first.
             if (_defaultParamSetEncountered)
             {
-                sb.AppendLine().Append('-', maxLineWidth - 1).AppendLine();
                 sb.Append(GetHelp(maxLineWidth, String.Empty, false));
             }
 
@@ -1587,7 +1634,6 @@ class CommandLine
             {
                 if (parameter.IsParameterSet && parameter.Name.Length != 0)
                 {
-                    sb.AppendLine().Append('-', maxLineWidth - 1).AppendLine();
                     sb.Append(GetHelp(maxLineWidth, parameter.Name, false));
                 }
             }
@@ -1595,8 +1641,8 @@ class CommandLine
             string globalQualifiers = GetHelpGlobalQualifiers(maxLineWidth);
             if (globalQualifiers.Length > 0)
             {
-                sb.Append('-', maxLineWidth - 1).AppendLine();
-                sb.Append("Qualifiers global to all commands:").AppendLine();
+                sb.AppendLine().Append('-', maxLineWidth - 1).AppendLine();
+                sb.Append("Global settings to all commands:").AppendLine();
                 sb.AppendLine();
                 sb.Append(globalQualifiers);
             }
@@ -1605,7 +1651,7 @@ class CommandLine
         }
         /// <summary>
         /// prints a string to the console in a nice way.  In particular it
-        /// displays a sceeen full of data and then as user to type a space to continue.
+        /// displays a screen full of data and then as user to type a space to continue.
         /// </summary>
         /// <param name="helpString"></param>
         private static void DisplayStringToConsole(string helpString)
@@ -1644,17 +1690,36 @@ class CommandLine
         private static void ParameterHelp(CommandLineParameter parameter, StringBuilder sb, int firstColumnWidth, int maxLineWidth)
         {
             // TODO alias information.
-            sb.Append("    ").Append(parameter.Syntax(true, true).PadRight(firstColumnWidth)).Append(' ');
+            sb.Append(parameter.Syntax().PadRight(firstColumnWidth));
             string helpText = parameter.HelpText;
-            if (typeof(Enum).IsAssignableFrom(parameter.Type))
-                helpText = helpText + "  Legal values: " + string.Join(", ", Enum.GetNames(parameter.Type)) + ".";
-            Wrap(sb, helpText, firstColumnWidth + 5, new string(' ', firstColumnWidth + 5), maxLineWidth);
+            string defValue = string.Empty;
+
+            bool shouldPrint = true;
+            // Bool type implied on named parameters
+            if (parameter.IsNamed && parameter.Type == typeof(bool))
+            {
+                shouldPrint = false;
+                if (parameter.Value != null && (bool)parameter.Value)
+                    shouldPrint = false;
+            }
+
+            if (shouldPrint)
+            {
+                if (!string.IsNullOrEmpty(parameter.DefaultValue))
+                {
+                    helpText += "\n=> Default value: " + parameter.DefaultValue.ToString();
+                }
+            }
+
+            if (parameter.LegalValues.Count > 0)
+                helpText = helpText + "\n=> Legal values: [" + string.Join(", ", parameter.LegalValues) + "].";
+            Wrap(sb, helpText, firstColumnWidth, new string(' ', firstColumnWidth), maxLineWidth, true);
         }
-        private static void Wrap(StringBuilder sb, string text, int startColumn, string linePrefix, int maxLineWidth)
+        private static void Wrap(StringBuilder sb, string text, int startColumn, string linePrefix, int maxLineWidth, bool first)
         {
             if (text != null)
             {
-                bool first = true;
+                //bool first = true;
                 int column = startColumn;
                 string previousWord = String.Empty;
                 string[] paragraphs = text.Split(new char[] { '\n' }, StringSplitOptions.RemoveEmptyEntries);
@@ -1662,9 +1727,10 @@ class CommandLine
                 {
                     if (!first)
                     {
-                        sb.AppendLine().AppendLine();
                         column = 0;
+                        linePrefix = new string(' ', startColumn);
                     }
+                    
                     string[] words = paragraph.Split((char[])null, StringSplitOptions.RemoveEmptyEntries);        // Split on whitespace
                     foreach (string word in words)
                     {
@@ -1694,6 +1760,7 @@ class CommandLine
                         column += word.Length;
                         first = false;
                     }
+                    sb.AppendLine();
                 }
             }
             sb.AppendLine();
@@ -1711,37 +1778,24 @@ class CommandLine
                 if (parameter.IsParameterSet)
                     break;
                 if (parameter.IsNamed)
-                    ParameterHelp(parameter, sb, QualifierSyntaxWidth, maxLineWidth);
+                    ParameterHelp(parameter, sb, GlobalQualifierSyntaxWidth, maxLineWidth);
             }
             return sb.ToString();
         }
 
         private static int GetConsoleWidth()
         {
-#if COREFX
             return 120; // Can't retrieve console width in .NET Core
-#else
-            return Console.WindowWidth;
-#endif
         }
 
         private static int GetConsoleHeight()
         {
-#if COREFX
-            return 40; // Can't retrieve console height in .NET Core
-#else
-            return Console.WindowHeight;
-#endif
+            return 100; // Can't retrieve console height in .NET Core
         }
 
         private static string GetEntryAssemblyName()
         {
-#if COREFX
-            // Cannot use Assembly.GetEntryAssembly(). This assumes that CommandLine is compiled into your application.
-            Assembly entryAssembly = typeof(CommandLineParser).GetTypeInfo().Assembly;
-#else
             Assembly entryAssembly = Assembly.GetEntryAssembly();
-#endif
             if (entryAssembly != null)
             {
                 return Path.GetFileNameWithoutExtension(entryAssembly.ManifestModule.Name);
@@ -1755,17 +1809,19 @@ class CommandLine
         // TODO expose the ability to change this?
         private static char[] s_separators = new char[] { ':', '=' };
 
-        // tweeks the user can specify
-        private bool _noDashOnParameterSets;
+        // tweaks the user can specify
+        private bool _noDashOnParameterSets = true;
         private bool _noSpaceOnQualifierValues;
         private string[] _parameterSetsWhereQualifiersMustBeFirst;
         private bool _qualifiersUseOnlyDash = true;
         private bool _lastQualifierWins;
+        private static string _extraParameters;
 
         // In order to produce help, we need to remember everything useful about all the parameters.  This list
         // does this.  It is only done when help is needed, so it is not here in the common scenario.
         private List<CommandLineParameter> _parameterDescriptions;
         private int _qualifierSyntaxWidth;       // When printing help, how much indent any line wraps.
+        private int _globalQualifierSyntaxWidth;
         public int QualifierSyntaxWidth
         {
             get
@@ -1777,15 +1833,48 @@ class CommandLine
                     if (_parameterDescriptions != null)
                     {
                         int maxSyntaxWidth = 0;
+                        bool qualifierStart = false;
                         foreach (CommandLineParameter parameter in _parameterDescriptions)
-                            maxSyntaxWidth = Math.Max(maxSyntaxWidth, parameter.Syntax(true, true).Length);
-                        _qualifierSyntaxWidth = Math.Max(8, maxSyntaxWidth + 1); // +1 leaves an extra space
+                        {
+                            if (parameter.IsParameterSet)
+                            {
+                                qualifierStart = true;
+                            }
+                            if(qualifierStart)
+                            {
+                                maxSyntaxWidth = Math.Max(maxSyntaxWidth, parameter.Syntax().Length);
+                            }
+                        }
+                        _qualifierSyntaxWidth = Math.Max(8, maxSyntaxWidth + 2); // +2 leaves two extra spaces
                     }
                 }
                 return _qualifierSyntaxWidth;
             }
         }
 
+        public int GlobalQualifierSyntaxWidth
+        {
+            get
+            {
+                // Find the optimal size for the 'first column' of the help text.
+                if (_globalQualifierSyntaxWidth == 0)
+                {
+                    _globalQualifierSyntaxWidth = 35;
+                    if (_parameterDescriptions != null)
+                    {
+                        int maxSyntaxWidth = 0;
+                        foreach (CommandLineParameter parameter in _parameterDescriptions)
+                        {
+                            if (parameter.IsParameterSet)
+                                break;
+                            maxSyntaxWidth = Math.Max(maxSyntaxWidth, parameter.Syntax().Length);
+                        }
+                        _globalQualifierSyntaxWidth = Math.Max(8, maxSyntaxWidth + 2); // +2 leaves two extra spaces
+                    }
+                }
+                return _globalQualifierSyntaxWidth;
+            }
+        }
 
         /// <summary>
         /// Qualifiers can have aliases (e.g. for short names).  This holds these aliases.
@@ -1800,7 +1889,7 @@ class CommandLine
         private Dictionary<string, int> _dashedParameterEncodedPositions;
         // We steal the top bit to prepresent whether the parameter occurs more than once.
         private const int MultiplePositions = unchecked((int)0x80000000);
-        private static int SetMulitple(int encodedPos) { return encodedPos | MultiplePositions; }
+        private static int SetMultiple(int encodedPos) { return encodedPos | MultiplePositions; }
         private static int GetPosition(int encodedPos) { return encodedPos & ~MultiplePositions; }
         private static bool IsMulitple(int encodedPos) { return (encodedPos & MultiplePositions) != 0; }
 
@@ -1808,6 +1897,7 @@ class CommandLine
         // ultimately ends up only having positional parameters being non-null.
         private List<string> _args;
 
+        private Setup _setupContent;
         private int _curPosition;                    // All arguments before this position have been processed.
         private bool _skipParameterSets;             // Have we found the parameter set qualifer, so we don't look at any others.
         private bool _skipDefinitions;               // Should we skip all subsequent definitions (typically until the next parameter set def)
